Cleanup and refactor edit.js code.
