+++ src/blocks/form-field/style.scss
@@ -4,8 +4,7 @@
 	gap: 0.5rem;
 
 	&.gatherpress-field-type-checkbox {
-		flex-direction: row-reverse;
-		justify-content: flex-end;
+		flex-direction: row;
 		align-items: center;
 		gap: 0.75rem;
